+++ test/StkTableMultiHeader.vue
@@ -40,7 +40,7 @@ const columns = [
         ],
     },
 ];
-const dataSource = new Array(20).fill(0).map((it, i) => {
+const dataSource = new Array(50).fill(0).map((it, i) => {
     return {
         id: i,
         lv2_1: 'lv2.1',
@@ -53,6 +53,6 @@ const dataSource = new Array(20).fill(0).map((it, i) => {
 <template>
     <div>
         多级表头处理
-        <StkTable style="height: 200px; width: 600px" fixed-col-shadow :columns="columns" :data-source="dataSource"> </StkTable>
+        <StkTable style="height: 200px; width: 600px" fixedColShadow virtual :columns="columns" :data-source="dataSource"> </StkTable>
     </div>
 </template>